demos: tour=async showcase; download_model clicks Chat mode + faster; vision 24x; settings no cursor detour; README refreshed
